repository test

+++ server/src/main/java/cz/pycrs/cloudquery/dto/MeasurementPatchRequest.java
@@ -11,6 +11,6 @@ public record MeasurementPatchRequest(
         Double maxTemperature,
         Double pressureSeaLevel,
         Double pressureGroundLevel,
-        Double humidity
+        Integer humidity
 ) {
 }
